requirements.txt
feedparser
python-dotenv
openai
requests
beautifulsoup4
fastapi
uvicorn
gspread
oauth2client
